+++ pkgs/jnigen/bin/download_maven_jars.dart
@@ -5,6 +5,7 @@
 import 'dart:io';
 
 import 'package:jnigen/jnigen.dart';
+import 'package:jnigen/src/tools/gradle_tools.dart';
 import 'package:jnigen/tools.dart';
 
 /// Downloads maven dependencies downloaded by equivalent jnigen invocation.
@@ -14,8 +15,13 @@ void main(List<String> args) async {
   final config = Config.parseArgs(args);
   final mavenDownloads = config.mavenDownloads;
   if (mavenDownloads != null) {
-    await MavenTools.downloadMavenJars(
-        MavenTools.deps(mavenDownloads.jarOnlyDeps + mavenDownloads.sourceDeps),
+    await GradleTools.downloadMavenSources(
+        GradleTools.deps(mavenDownloads.sourceDeps), mavenDownloads.sourceDir);
+    // Include sources in jar download to make sure transitive
+    // dependencies are included
+    await GradleTools.downloadMavenJars(
+        GradleTools.deps(
+            mavenDownloads.jarOnlyDeps + mavenDownloads.sourceDeps),
         mavenDownloads.jarDir);
     await Directory(mavenDownloads.jarDir)
         .list()