+++ pkgs/jnigen/lib/src/util/find_package.dart
@@ -37,10 +37,22 @@ Future<bool> isPackageModifiedAfter(String packageName, DateTime time,
   // In case of git / pub package we might be able to check pubspec, but no
   // such technique applies for path packages.
   await for (final entry in dir.list(recursive: true)) {
-    final stat = await entry.stat();
-    if (stat.modified.isAfter(time)) {
-      return true;
+    if (!isExcludedBuildPath(entry)) {
+      final stat = await entry.stat();
+      if (stat.modified.isAfter(time)) {
+        print(stat);
+        print(entry);
+        return true;
+      }
     }
   }
   return false;
+}
+
+bool isExcludedBuildPath(FileSystemEntity entry) {
+  return entry.path.contains('pom.xml') ||
+      entry.path.endsWith('.md') ||
+      entry.path.contains('.gradle') ||
+      entry.path.contains('.idea') ||
+      entry.path.contains('build');
 }