+++ pkgs/jnigen/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.14.1-wip
 - Added support for generating matching Kotlin operators as Dart operators.
+- Retrieval of dependencies uses Gradle (`GradleTools`) en lieu of Maven (`MavenTools`).
+- Updated `bin\download_maven_jars.dart` to use `GradleTools`.
 
 ## 0.14.0